Add a little bit more output.
There isn't much output from the MCMC, so there's the luxury to increase
the verbosity.  This now prints notices every time the actual covariance is
overridden by a user value.  It also prints the trace and diagonal elements
of the covariance of the accepted points every time the proposal
distribution is updated.