Accumulated changes to save the state and restart a chain
This implements the changes necessary to save the state of a chain in
the output file, and to restart the change from a previous file.  It
also implements the ability to provide correlations between variables
when setting the initial proposal covariance.

These are changes that appear to have been made mostly done in 2018, but
were then forgotten.  I've finished the changes and done the testing now.