Fix the target acceptance rate, and add citation
Shilin pointed out that the target acceptance rate at high dimensions
should be have an upper limit of 23.4%, but I had a hard coded value
of 44% (which is for one dimension.  I've changed it so you can set
the acceptance rate from the main program, and added default values
set based on the number of dimensions.  I also added a citation to a
detailed discussion, but it essentially says "finding the optimal
acceptance rate is hard".