When covariance is not positive definite use eigenvalues
When the covariance is very problematic, this usually means that it's
degenerate in one or more dimensions.  Try to use eigenvalue
decomposition to remove degeneracies, and hope for the best.  The
place I have seen this happen is when the user requests a high
dimensional proposed step that is almost totally correlated in a few
dimensions.