New Feature: Scan a dimension of the likelihood
This adjust the TProposeAdaptiveStep and TSimpleMCMC class/template to
allow the likelihood versus the input variables to be calculated and
saved.  This allows plots of Likelihood vs variable to be made.

TProposeAdaptiveStep : Add SetScanDimension(int) which will cause the
input dimension to be varied around the current central point.  The
new steps will be distributed according to the variation defined in
ProposalType.

TSimpleMCMC : Update the Step method so that the "metropolis" argument
is now an integer Values: (0) use a metropolis step [the default]' (1)
use a downhill step; (2) Accept all steps (used for scanning a
likelihood dimension.